requirements/base.txt
amqp==5.2.0; python_version >= '3.6'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
billiard==4.2.0; python_version >= '3.7'
celery==5.4.0; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
cloudinary==1.41.0
cron-descriptor==1.4.3
cryptography==43.0.0; python_version >= '3.7'
defusedxml==0.8.0rc2; python_version >= '3.6'
django==5.0.7; python_version >= '3.10'
django-appconf==1.0.6; python_version >= '3.7'
django-autoslug==1.9.9
django-celery-beat==2.6.0
django-celery-email==3.0.0
django-countries==7.6.1
django-filter==24.3; python_version >= '3.8'
django-phonenumber-field==8.0.0; python_version >= '3.8'
django-taggit==6.0.0; python_version >= '3.8'
django-templated-mail==1.1.1
django-timezone-field==7.0; python_version >= '3.8' and python_version < '4.0'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djoser==2.2.3; python_version >= '3.8' and python_version < '4.0'
drf-yasg==1.21.7; python_version >= '3.6'
flower==2.0.1; python_version >= '3.7'
humanize==4.10.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
kombu==5.3.7; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
phonenumbers==8.13.42
pillow==10.4.0; python_version >= '3.8'
prometheus-client==0.20.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.9.0; python_version >= '3.8'
python-crontab==3.2.0
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python3-openid==3.2.0
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
redis==5.0.8; python_version >= '3.7'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.4.2; python_version >= '3.8'
social-auth-core==4.5.4; python_version >= '3.8'
sqlparse==0.5.1; python_version >= '3.8'
tornado==6.4.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.2; python_version >= '3.8'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13